requirements.txt
flask
psycopg2-binary
bcrypt
pyjwt
python-dotenv
